+++ honbu/deps.ts
@@ -1,10 +1,17 @@
+import "https://deno.land/std@0.160.0/dotenv/load.ts";
 export {
   parse,
   stringify,
 } from "https://deno.land/std@0.158.0/encoding/yaml.ts";
 export { config } from "https://deno.land/x/dotenv@v3.2.0/mod.ts";
-export { getNetworkAddr } from "https://deno.land/x/local_ip@0.0.3/mod.ts";
 export * as oak from "https://deno.land/x/oak@v11.1.0/mod.ts";
 export { oakCors } from "https://deno.land/x/cors@v1.2.2/mod.ts";
 export { isString } from "https://deno.land/std@0.158.0/encoding/_yaml/utils.ts";
 export { Cron } from "https://deno.land/x/croner@5.3.0/src/croner.js";
+export * as mongodb from "https://deno.land/x/mongo@v0.31.1/mod.ts";
+export { Mutex } from "https://deno.land/x/semaphore@v1.1.1/mod.ts";
+export * as base64 from "https://deno.land/std@0.160.0/encoding/base64url.ts";
+export * as log from "https://deno.land/std@0.160.0/log/mod.ts";
+export * as bcrypt from "https://deno.land/x/bcrypt@v0.4.1/mod.ts";
+export * as djwt from "https://deno.land/x/djwt@v2.7/mod.ts";
+export * from "https://deno.land/x/fido2@3.3.4/dist/main.js";
