+++ honbu/ebinaAPI/app/api.ts
@@ -1,19 +1,30 @@
 import { oak } from "../../deps.ts";
-import { appURL } from "../ebina.ts";
+import { authToken } from "../../utils/auth.ts";
+import { APIs } from "../../settings/apis.ts";
+import { APPS_DIR } from "../../settings/settings.ts";
 
 const apiRouter = new oak.Router();
 
+const entrances: {
+  [name: string]: {
+    entranceProc: Deno.Process | null;
+    startedDate: number | null;
+  } | undefined;
+} = {};
+
+const apisList: { [name: string]: APIs | undefined } = {};
+
 // API起動状態取得
 // 200 { status: 'started' | 'stop', started_at: number }
-apiRouter.get("/status", async (ctx) => {
+apiRouter.get("/status", authToken, (ctx) => {
   const { appName } = ctx.params;
-  await fetch(`${appURL}/${appName}/api/status`, {
-    method: "GET",
-    headers: ctx.request.headers,
-  }).then(async (ret) => {
-    ctx.response.body = await ret.json();
-    ctx.response.status = ret.status;
-  });
+  if (!appName) return ctx.response.status = 400;
+  const entrance = entrances[appName];
+  const isStarted = entrance && entrance.entranceProc;
+  ctx.response.body = {
+    status: isStarted ? "started" : "stop",
+    started_at: isStarted ? entrance.startedDate : undefined,
+  };
 });
 
 // API起動状態更新
@@ -21,58 +32,82 @@ apiRouter.get("/status", async (ctx) => {
 // 200 できた
 // 400 情報おかしい
 // 500 起動できなかった
-apiRouter.put("/status", async (ctx) => {
+apiRouter.put("/status", authToken, async (ctx) => {
   const { appName } = ctx.params;
-  await fetch(`${appURL}/${appName}/api/status`, {
-    method: "PUT",
-    headers: ctx.request.headers,
-    body: await ctx.request.body({ type: "text" }).value,
-  }).then(async (ret) => {
-    ctx.response.body = await ret.json();
-    ctx.response.status = ret.status;
-  });
+  if (!appName) return ctx.response.status = 400;
+  const { status } = await ctx.request.body({ type: "json" }).value;
+  if (!status) return ctx.response.status = 400;
+  const isStop = status === "stop";
+  const entrance = entrances[appName];
+
+  if (entrance && entrance.entranceProc) {
+    entrance.entranceProc.kill("SIGINT");
+    entrance.entranceProc = null;
+    entrance.startedDate = null;
+    if (isStop) return ctx.response.body = { message: "stop" };
+  } else if (isStop) return ctx.response.body = { message: "already stoped" };
+
+  switch (status) {
+    case "start": {
+      const entranceProc = Deno.run({
+        cmd: [
+          "deno",
+          "run",
+          "--allow-all",
+          "./entrance.ts",
+          `${APPS_DIR}/${appName}`,
+        ],
+      });
+      const startedDate = Date.now();
+
+      entrances[appName] = { entranceProc, startedDate };
+      ctx.response.body = { message: "start" };
+      break;
+    }
+
+    default:
+      ctx.response.status = 400;
+      break;
+  }
 });
 
 // ポート取得
 // 200 { port: number }
-apiRouter.get("/port", async (ctx) => {
+apiRouter.get("/port", authToken, (ctx) => {
   const { appName } = ctx.params;
-  await fetch(`${appURL}/${appName}/api/port`, {
-    method: "GET",
-    headers: ctx.request.headers,
-  }).then(async (ret) => {
-    ctx.response.body = await ret.json();
-    ctx.response.status = ret.status;
-  });
+  if (!appName) return ctx.response.status = 400;
+  const apisInstance = apisList[appName] ??
+    (apisList[appName] = new APIs(appName));
+  ctx.response.body = { port: apisInstance.getPort() };
 });
 
 // ポート設定
 // { port: number }
 // 200 OK
 // 400 情報おかしい
-apiRouter.put("/port", async (ctx) => {
+apiRouter.put("/port", authToken, async (ctx) => {
   const { appName } = ctx.params;
-  await fetch(`${appURL}/${appName}/api/port`, {
-    method: "PUT",
-    headers: ctx.request.headers,
-    body: await ctx.request.body({ type: "text" }).value,
-  }).then(async (ret) => {
-    ctx.response.body = await ret.json();
-    ctx.response.status = ret.status;
-  });
+  if (!appName) return ctx.response.status = 400;
+  const { port } = await ctx.request.body({ type: "json" }).value;
+  if (!port) return ctx.response.status = 400;
+
+  const apisInstance = apisList[appName] ??
+    (apisList[appName] = new APIs(appName));
+  apisInstance.setPort(port);
+  ctx.response.status = 200;
 });
 
 // API一覧取得
 // 200 { path, api }
-apiRouter.get("/endpoint", async (ctx) => {
+apiRouter.get("/endpoint", authToken, (ctx) => {
   const { appName } = ctx.params;
-  await fetch(`${appURL}/${appName}/api/endpoint`, {
-    method: "GET",
-    headers: ctx.request.headers,
-  }).then(async (ret) => {
-    ctx.response.body = await ret.json();
-    ctx.response.status = ret.status;
-  });
+  if (!appName) return ctx.response.status = 400;
+  const apisInstance = apisList[appName] ??
+    (apisList[appName] = new APIs(appName));
+
+  const apis = apisInstance.getAPIs();
+  const apiList = Object.keys(apis).map((path) => ({ path, api: apis[path] }));
+  ctx.response.body = apiList;
 });
 
 // API作成
@@ -80,16 +115,19 @@ apiRouter.get("/endpoint", async (ctx) => {
 // { name, method, type, value }
 // 200 OK
 // 400 情報おかしい
-apiRouter.post("/endpoint/:path", async (ctx) => {
+apiRouter.post("/endpoint/:path", authToken, async (ctx) => {
   const { appName, path } = ctx.params;
-  await fetch(`${appURL}/${appName}/api/endpoint/${path}`, {
-    method: "POST",
-    headers: ctx.request.headers,
-    body: await ctx.request.body({ type: "text" }).value,
-  }).then(async (ret) => {
-    ctx.response.body = await ret.json();
-    ctx.response.status = ret.status;
-  });
+  if (!appName) return ctx.response.status = 400;
+  const { name, method, type, value } = await ctx.request.body({ type: "json" })
+    .value;
+  if (!path || !name || !method || !type || !value) {
+    return ctx.response.status = 400;
+  }
+
+  const apisInstance = apisList[appName] ??
+    (apisList[appName] = new APIs(appName));
+  apisInstance.setAPI(path, { name, method, type, value });
+  ctx.response.status = 200;
 });
 
 // API取得
@@ -97,47 +135,50 @@ apiRouter.post("/endpoint/:path", async (ctx) => {
 // 200 API
 // 400 情報おかしい
 // 404 ない
-apiRouter.get("/endpoint/:path", async (ctx) => {
+apiRouter.get("/endpoint/:path", authToken, (ctx) => {
   const { appName, path } = ctx.params;
-  await fetch(`${appURL}/${appName}/api/endpoint/${path}`, {
-    method: "GET",
-    headers: ctx.request.headers,
-  }).then(async (ret) => {
-    ctx.response.body = await ret.json();
-    ctx.response.status = ret.status;
-  });
+  if (!appName) return ctx.response.status = 400;
+
+  const apisInstance = apisList[appName] ??
+    (apisList[appName] = new APIs(appName));
+  const api = apisInstance.getAPI(path);
+  if (api) {
+    ctx.response.body = api;
+  } else {
+    ctx.response.status = 404;
+  }
 });
 
 // API更新
 // :path
 // 200 OK
 // 400 情報おかしい
-apiRouter.put("/endpoint/:path", async (ctx) => {
+apiRouter.put("/endpoint/:path", authToken, async (ctx) => {
   const { appName, path } = ctx.params;
-  await fetch(`${appURL}/${appName}/api/endpoint/${path}`, {
-    method: "PUT",
-    headers: ctx.request.headers,
-    body: await ctx.request.body({ type: "text" }).value,
-  }).then(async (ret) => {
-    ctx.response.body = await ret.json();
-    ctx.response.status = ret.status;
-  });
+  if (!appName) return ctx.response.status = 400;
+  const { name, method, type, value } = await ctx.request.body({ type: "json" })
+    .value;
+  if (!path || !name || !method || !type || !value) {
+    return ctx.response.status = 400;
+  }
+
+  const apisInstance = apisList[appName] ??
+    (apisList[appName] = new APIs(appName));
+  apisInstance.setAPI(path, { name, method, type, value });
+  ctx.response.status = 200;
 });
 
 // API削除
 // 200 OK
 // 400 情報おかしい
 // 404 パスない
-apiRouter.delete("/endpoint/:path", async (ctx) => {
+apiRouter.delete("/endpoint/:path", authToken, (ctx) => {
   const { appName, path } = ctx.params;
-  await fetch(`${appURL}/${appName}/api/endpoint/${path}`, {
-    method: "DELETE",
-    headers: ctx.request.headers,
-    body: await ctx.request.body({ type: "text" }).value,
-  }).then(async (ret) => {
-    ctx.response.body = await ret.json();
-    ctx.response.status = ret.status;
-  });
+  if (!appName) return ctx.response.status = 400;
+
+  const apisInstance = apisList[appName] ??
+    (apisList[appName] = new APIs(appName));
+  ctx.response.status = apisInstance.deleteAPI(path) ? 200 : 404;
 });
 
 export default apiRouter;
