+++ honbu/utils/webauthn/funcs.ts
