+++ initEbinaStation.sh
@@ -7,9 +7,6 @@ if !(type deno > /dev/null 2>&1); then
   exit 1
 fi
 
-echo "Compile EbinaStation Honbu"
-deno compile --allow-run --allow-read --allow-write --allow-net --allow-env -o ./generate/ebinaStationHonbu ./honbu/main.ts
-
 if (type systemctl > /dev/null 2>&1); then
   sed -e 's/ExecStart.*/ExecStart=\/app\/ebina-station\/startEbinaStation.sh/' ebina-station.service.base > ./generate/ebina-station.service
   echo "link ebina-station.service to systemd"
