--- koujou/Dockerfile
@@ -1,15 +1,0 @@
-FROM denoland/deno:latest
-
-EXPOSE 3456
-
-WORKDIR /app
-
-COPY deps.ts .
-RUN deno cache deps.ts
-
-COPY . .
-
-RUN mkdir logs
-RUN deno cache index.ts
-
-CMD ["run", "-A", "index.ts"]
