--- koujou/project
@@ -1,1 +1,0 @@
-../project
