+++ startEbinaStation.sh
@@ -2,7 +2,7 @@
 cd `dirname $0`
 
 if [ -e ./generate/ebinaStationHonbu ]; then
-  ./generate/ebinaStationHonbu
+  deno run -A ./honbu/main.ts
 else
   echo "init first"
   exit 1
